Add command 'delete user'
