+++ cmd/delete.go
@@ -23,8 +23,18 @@ var deleteMcpClientCmd = &cobra.Command{
 	RunE: runDeleteMcpClient,
 }
 
+var deleteUserCmd = &cobra.Command{
+	Use:   "user [username]",
+	Args:  cobra.ExactArgs(1),
+	Short: "Delete a user (Production mode)",
+	Long:  "Delete a user from mcpjungle.\nThis instantly revokes all access of this user.",
+	RunE:  runDeleteUser,
+}
+
 func init() {
 	deleteCmd.AddCommand(deleteMcpClientCmd)
+	deleteCmd.AddCommand(deleteUserCmd)
+
 	rootCmd.AddCommand(deleteCmd)
 }
 
@@ -34,5 +44,14 @@ func runDeleteMcpClient(cmd *cobra.Command, args []string) error {
 		return fmt.Errorf("failed to delete the client: %w", err)
 	}
 	fmt.Printf("MCP client '%s' deleted successfully (if it existed)!\n", name)
+	return nil
+}
+
+func runDeleteUser(cmd *cobra.Command, args []string) error {
+	username := args[0]
+	if err := apiClient.DeleteUser(username); err != nil {
+		return fmt.Errorf("failed to delete the user: %w", err)
+	}
+	cmd.Printf("User '%s' deleted successfully (if they existed)\n", username)
 	return nil
 }
